Refactored component folder names and Drawer into SidePanel component
