+++ frontend/src/components/editor/RichTextEditor.tsx
