+++ frontend/src/components/editor/menu/MenuBar.tsx
