+++ frontend/src/components/editor/styles.css
