+++ frontend/src/components/logo/Logo.tsx
